requirements.txt
Django>=2.2.1, <2.3.0
djangorestframework>=3.9.4, <3.10.0
djangorestframework-jwt>=1.11.0,<1.12.0
psycopg2>=2.8.2,<2.9.0
djoser>=1.7,<1.80
flake8>=3.7.7, <3.8.0
django-rest-swagger